add TranslatorGetResponse, UserGetResponse

+++ src/main/java/com/example/morago/model/dto/requests/user/UserProfileUpdateRequest.java
@@ -3,6 +3,7 @@ package com.example.morago.model.dto.requests.user;
 import com.example.morago.model.dto.requests.UserUpdateRequest;
 import com.example.morago.model.entity.UserProfile;
 import io.swagger.v3.oas.annotations.media.Schema;
+import java.math.BigDecimal;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -12,7 +13,7 @@ import lombok.Setter;
 public class UserProfileUpdateRequest extends UserUpdateRequest {
 
     @Schema(description = "Coins", example = "1000")
-    private Long coins;
+    private BigDecimal coins;
 
     public UserProfile build(UserProfile userProfile)
     {
